fix: port libjxl safety fixes and tune effort parameters
- effort.rs: doubled max_buckets at all efforts and added effort-scaled
  sampling fractions matching libjxl PR #4236 (e5=0.25, e7=0.5, e8=0.55)
- predictor.rs: use i64 intermediates in Average4 to prevent overflow
  (matches libjxl PR #4574)
- patches.rs: safe_round_to_i32/safe_trunc_to_i32 to prevent float-to-int
  overflow on extreme values (matches libjxl PR #4596)
- encode.rs: add use_squeeze parameter for DC frame Squeeze support,
  refactor GroupHeader to handle both RCT and Squeeze transforms
- lf_frame.rs: explicitly pass use_squeeze=false (pre-quantized DC data
  compresses worse with Squeeze; proper fix needs lossy modular quantization)
- CLAUDE.md: correct LfFrame overhead root cause (28% gap from architectural
  difference: responsive=1/Squeeze/lossy-modular vs pre-quantized/lossless)

+++ jxl_encoder/src/modular/predictor.rs
@@ -115,7 +115,14 @@ impl Predictor {
             Predictor::Average4 => {
                 // AverageAll: (6*N - 2*NN + 7*W + WW + NEE + 3*NE + 8) / 16
                 // where NEE = toprightright = pixel at (x+2, y-1)
-                (6 * n.n - 2 * n.nn + 7 * n.w + n.ww + n.nee + 3 * n.ne + 8) / 16
+                // Use i64 intermediates to prevent overflow (libjxl PR #4574)
+                ((6i64 * n.n as i64 - 2 * n.nn as i64
+                    + 7 * n.w as i64
+                    + n.ww as i64
+                    + n.nee as i64
+                    + 3 * n.ne as i64
+                    + 8)
+                    / 16) as i32
             }
         }
     }
